requirements.txt
setuptools
graphviz
duckdb
google-generativeai
grpcio==1.67.1
watchdog
pandas
streamlit
json5
groq